backend/src/requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.6
cryptography==3.4.8
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic-core==2.3.0
pydantic-settings==2.0.2
PyJWT==1.7.1
pytest==6.2.5
pytest-pythonpath==0.7.4
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
trio==0.21.0
typing-extensions==4.7.1
uvicorn==0.23.1
Werkzeug==2.3.6
zipp==3.16.2
